feedback 103, 105,110, 111, 121, 122

+++ packages/frontend/src/components/views/ganntView/GanttView.tsx
@@ -432,7 +432,7 @@ function GanttView() {
         nonworkingDays={HandleNonWorkingDays()}
         monthScale={monthScale}
         monthFormat="long"
-        dayFormat="long"
+        dayFormat="numeric"
         weekFormat="long"
         disableSelection
         view={filterUnit}
